Working towards d3d+gl dual support again. Fixed slow-rockets bug with high framerates.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/branches/wip@3608 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/server/sv_phys.c
@@ -1836,7 +1836,7 @@ void SV_RunEntity (edict_t *ent)
 	}
 	else
 	{
-		if (ent->v->lastruntime == svs.framenum)
+		if ((unsigned int)ent->v->lastruntime == svs.framenum)
 			return;
 		ent->v->lastruntime = svs.framenum;
 	}
